+++ src/main/scala/sangria/marshalling/MarshallingUtil.scala
@@ -36,4 +36,45 @@ object MarshallingUtil {
   implicit class MarshaledConverter[In : InputUnmarshaller](in: In) {
     def convertMarshaled[Out : ResultMarshallerForType] = convert(in)
   }
+
+  implicit class ResultMarshallerOps(val m: ResultMarshaller) extends AnyVal {
+    def list(elements: ResultMarshaller#Node*): m.Node =
+      m.arrayNode(elements.asInstanceOf[Seq[m.Node]].toVector)
+
+    def map(elements: (String, ResultMarshaller#Node)*): m.Node =
+      m.mapNode(elements.foldLeft(m.emptyMapNode(elements.map(_._1))) {
+        case (acc, (name, value)) ⇒ m.addMapNodeElem(acc, name, value.asInstanceOf[m.Node], optional = false)
+      })
+
+    def fromString(value: String): m.Node =
+      m.scalarNode(value, "String", Set.empty)
+
+    def fromEnumString(value: String): m.Node =
+      m.enumNode(value, "")
+
+    def fromInt(value: Int): m.Node =
+      m.scalarNode(value, "Int", Set.empty)
+
+    def fromLong(value: Long): m.Node =
+      m.scalarNode(value, "Long", Set.empty)
+
+    def fromBoolean(value: Boolean): m.Node =
+      m.scalarNode(value, "Long", Set.empty)
+
+    def fromFloat(value: Float): m.Node =
+      m.scalarNode(value, "Float", Set.empty)
+
+    def fromDouble(value: Double): m.Node =
+      m.scalarNode(value, "Float", Set.empty)
+
+    def fromBigInt(value: BigInt): m.Node =
+      m.scalarNode(value, "BigInt", Set.empty)
+
+    def fromBigInt(value: BigDecimal): m.Node =
+      m.scalarNode(value, "BigDecimal", Set.empty)
+
+    def fromInput[Input : InputUnmarshaller](value: Input): m.Node = {
+      value.convertMarshaled(SimpleResultMarshallerForType[m.Node](m))
+    }
+  }
 }